backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0

aiofiles==24.1.0
dnspython==2.6.1

motor==3.6.0
pydantic==2.9.2
email-validator==2.2.0
python-dotenv==1.0.1
python-multipart==0.0.9

python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2

google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3

langdetect==1.0.9
textblob==0.18.0

twilio==9.3.2
requests==2.32.3

instaloader==4.13.1
stripe==9.8.0
paypalrestsdk==1.13.3

beautifulsoup4==4.12.3
cohere==5.9.0

razorpay==1.3.1
openai==1.8.0


redis>=5.0.0
pandas==2.2.3
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2


aiosmtplib

python-socketio
jinja2